autotranslated python2 files to python3

+++ python/ofdm/qa_fbmc_symbol_estimation_vcb.py
@@ -21,8 +21,8 @@
 
 from gnuradio import gr, gr_unittest
 from gnuradio import blocks
-from fbmc_symbol_estimation_vcb import fbmc_symbol_estimation_vcb
-from fbmc_symbol_creation_bvc import fbmc_symbol_creation_bvc
+from .fbmc_symbol_estimation_vcb import fbmc_symbol_estimation_vcb
+from .fbmc_symbol_creation_bvc import fbmc_symbol_creation_bvc
 
 import math
 import random
@@ -37,21 +37,21 @@ class qa_fbmc_fbmc_symbol_estimation_vcb (gr_unittest.TestCase):
 
     def test_001_t (self):
         # set up fg
-        src_data = (0,1,3,2,3,1,0,0,2,1,2,3)
+        src_data = (0, 1, 3, 2, 3, 1, 0, 0, 2, 1, 2, 3)
         expected_result = src_data
         src = blocks.vector_source_b(src_data)
         cre = fbmc_symbol_creation_bvc(M=4, qam_size=4)
-        est = fbmc_symbol_estimation_vcb(M=4,qam_size=4)
+        est = fbmc_symbol_estimation_vcb(M=4, qam_size=4)
         dst = blocks.vector_sink_b()
-        self.tb.connect(src,cre,est,dst)
+        self.tb.connect(src, cre, est, dst)
         self.tb.run ()
         # check data
         result_data = dst.data()
-        self.assertEqual(expected_result,result_data)
+        self.assertEqual(expected_result, result_data)
 
     def test_002_t (self):
-        M=int(math.pow(2,6))
-        num_per_sc = int(math.pow(2,16))
+        M=int(math.pow(2, 6))
+        num_per_sc = int(math.pow(2, 16))
         mapper = {0:0.7071067690849304+0.7071067690849304j,
         1:-0.7071067690849304+0.7071067690849304j,
         2:-0.7071067690849304-0.7071067690849304j,
@@ -61,17 +61,17 @@ class qa_fbmc_fbmc_symbol_estimation_vcb (gr_unittest.TestCase):
         for i in range(M*num_per_sc):
             randy = int(random.random()*10)%4
             src_data.append(mapper[randy])
-            expected_result.append(long(randy))
+            expected_result.append(int(randy))
 
-        src = blocks.vector_source_c(src_data,vlen=M)
-        est = fbmc_symbol_estimation_vcb(M=M,qam_size=4)
+        src = blocks.vector_source_c(src_data, vlen=M)
+        est = fbmc_symbol_estimation_vcb(M=M, qam_size=4)
         dst = blocks.vector_sink_b()
 
-        self.tb.connect(src,est,dst)
+        self.tb.connect(src, est, dst)
         self.tb.run ()
         # check data
         result_data = dst.data()
-        self.assertEqual(tuple(expected_result),result_data)
+        self.assertEqual(tuple(expected_result), result_data)
 
 
 if __name__ == '__main__':
